+++ tests/dummy/config/environment.js
@@ -1,10 +1,7 @@
 /* jshint node: true */
 
 module.exports = function(environment) {
-  // var backendUrl = 'https://flexberry-ember-dummy.azurewebsites.net';
-  var backendUrl = 'http://localhost:6500';
-
-  // var backendUrl = 'http://rtc-web.ics.perm.ru:8081';
+  var backendUrl = 'http://flexberry.northeurope.cloudapp.azure.com';
 
   if (environment === 'development-loc') {
     // Use `ember s -e development-loc` command for local backend usage.
@@ -131,6 +128,12 @@ module.exports = function(environment) {
     // Testem prefers this...
     ENV.baseURL = '/';
     ENV.locationType = 'none';
+
+    // URL of the backend running in docker.
+    backendUrl = 'http://localhost:6500';
+    ENV.APP.backendUrl = backendUrl;
+    ENV.APP.backendUrls.root = backendUrl;
+    ENV.APP.backendUrls.api = backendUrl + '/odata';
 
     // keep test console output quieter
     ENV.APP.LOG_ACTIVE_GENERATION = false;