+++ CHANGELOG.md
@@ -6,6 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 * `reload-list-mixin` mixin:
     * Fix error, when use DatePredicate in filter. Create only SimplePredicate from SimplePredicate and DatePredicate.
+    * Fix error, when use empty value in conditios "eq" and "like" by filters types data, string, number. 
+* `predicate-from-filters` mixin:
+    * Fix error, when use empty value in conditios "eq" and "like" by filters types data, string, number. 
 
 ## [2.1.0-beta.0] - 2018-12-13
 ### Added