+++ CHANGELOG.md
@@ -3,6 +3,10 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+* `flexberry-groupedit` component:
+    * Add text and icon custom buttons in toolbar.
+
 ### Fixed
 * `reload-list-mixin` mixin:
     * Fix error, when use DatePredicate in filter. Create only SimplePredicate from SimplePredicate and DatePredicate.